refactor(execution-engine): extract task-runner-pool
Slice 2 of the execution-engine god-file decomposition. Pool selection,
SSH leasing, and executor construction move from task-runner.ts into
task-runner-pool.ts as free functions over a Pick<TaskRunner, ...> host,
matching the task-runner-phase-host.ts convention. TaskRunner keeps thin
delegates; the public surface and behavior are unchanged.
