+++ packages/execution-engine/src/index.ts
@@ -25,6 +25,11 @@ export { bashNormalizeTildePath, execRemoteCapture, shellPosixSingleQuote } from
 export * from './repo-pool.js';
 export * from './registry.js';
 export * from './task-runner.js';
+export type {
+  ExecutionPoolMember,
+  ExecutionPoolConfig,
+  PoolSelection,
+} from './task-runner-pool.js';
 export * from './harness-capabilities.js';
 export * from './merge-runner.js';
 export * from './conflict-resolver.js';
